requirements.txt
astroid==2.0.4
autopep8==1.4
Django==1.11
djangorestframework==3.6.0
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
